Refactor: 분리된 컴포넌트(CenterModal, SubModal)를 합쳐서 하나의 컴포넌트(Modal) 생성
- position: 모달의 위치를 지정
- overlayDisabled: 모달 바깥 오버레이를 클릭하면 모달창이 꺼지는 것을 막는 불리언 속성
